hadolint: ignore DL3008 and DL4006
DL3008 is about version pinning when doing apt-get install. We get the
packages from a subshell (line 72), so this doesn't really make sense.
Besides, it is not clear to us if we want to pin these, as the apt
repositories are very stable and only have one version of each package
available for download.

DL4006 mandates that we set -o pipefail because of the pipe on line 77,
but this doesn't work as -o pipefail is undefined in POSIX sh (SC2039).